+++ futurehome2mqtt/pyfimptoha/homeassistant.py
@@ -2,6 +2,7 @@ import json
 import time
 import paho.mqtt.client as client
 import pyfimptoha.sensor as sensor
+import pyfimptoha.light as light
 
 
 def create_components(
@@ -18,8 +19,6 @@ def create_components(
     print('Devices without rooms are ignored')
 
     statuses = []
-
-    # todo Add support for Wall plugs with functionality 'Lighting'
 
     for device in devices:
         address = device["fimp"]["address"]
@@ -41,8 +40,8 @@ def create_components(
 
         for service_name, service in device["services"].items():
             # Adding sensors
-            # todo add more sensors: alarm_power?, sensor_power. see sensor.py.01
-            status = False
+            # todo add more sensors: alarm_power?, sensor_power. see old sensor.py
+            status = None
             if service_name == "battery":
                 print(f"- Service: {service_name}")
                 status = sensor.battery(
@@ -74,73 +73,26 @@ def create_components(
                 statuses.append(status)
 
             # Lights
-            # todo Move to separate file
             elif functionality == "lighting":
+                status = None
                 if service_name == "out_lvl_switch":
-                    identifier = f"fh_{address}_{service_name}"
-                    command_topic = f"pt:j1/mt:cmd/rt:dev/rn:zw/ad:1/sv:out_lvl_switch/ad:{address}_1"
-                    state_topic = f"pt:j1/mt:evt/rt:dev/rn:zw/ad:1/sv:out_lvl_switch/ad:{address}_1"
-                    component = {
-                        "name": f"{device['client']['name']}",
-                        "object_id": identifier,
-                        "unique_id": identifier,
-                        "schema": "template",
-                        "command_topic": command_topic,
-                        "state_topic": state_topic,
-                        "brightness_scale": 100,
-                        "command_on_template": """
-                            {
-                              "props":{},
-                              "serv":"out_lvl_switch",
-                              "tags":[]
-                            {%- if brightness -%}
-                              , "type":"cmd.lvl.set",
-                              "val":{{ (brightness / 2.55) | int }},
-                              "val_t":"int"
-                            {%- else -%}
-                              , "type":"cmd.binary.set",
-                              "val":true,
-                              "val_t":"bool"
-                            {%- endif -%}
-                            }
-                        """,
-                        "command_off_template": """
-                            {"props":{},
-                              "serv":"out_lvl_switch",
-                              "tags":[],
-                              "type":"cmd.binary.set",
-                              "val":false,
-                              "val_t":"bool"}
-                        """,
-                        "state_template": "{% if value_json.val %} on {% else %} off {% endif %}",
-                        "brightness_template": "{% if value_json.val_t %}{{ (value_json.val * 2.55) | int }}{% endif %}"
-                    }
-                    payload = json.dumps(component)
-                    mqtt.publish(f"homeassistant/light/{identifier}/config", payload)
+                    print(f"- Service: {service_name}")
+                    status = light.out_lvl_switch(
+                        service_name=service_name,
+                        device=device,
+                        mqtt=mqtt,
+                        service=service,
+                    )
+                elif service_name == "out_bin_switch":
+                    status = light.out_bin_switch(
+                        service_name=service_name,
+                        device=device,
+                        mqtt=mqtt,
+                        service=service,
+                    )
 
-                    # Push statuses
-                    # todo binary switches are not set correctly
-                    if device.get("param") and device['param'].get('power'):
-                        power = device['param']['power']
-                        if power == "off":
-                            data = {
-                                "props": {},
-                                "serv": "out_lvl_switch",
-                                "type": "cmd.binary.report",
-                                "val_t": "bool",
-                                "val": False
-                            }
-                        else:
-                            dim_value = device['param']['dimValue']
-                            data = {
-                                "props": {},
-                                "serv": "out_lvl_switch",
-                                "type": "cmd.lvl.report",
-                                "val_t": "int",
-                                "val": dim_value
-                            }
-                        payload = json.dumps(data)
-                        statuses.append((state_topic, payload))
+                if status:
+                    statuses.append(status)
 
             # Appliance
             elif functionality == "appliance":